requirements.txt
pywebview~=3.4
pytz~=2021.1
selenium~=3.141.0